pyproject.toml
[project]
name = "crop-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=23.0.0",
    "boto3>=1.34.0",
    "faiss-cpu>=1.8.0",
    "fastapi>=0.110.0",
    "flake8>=6.0.0",
    "google-generativeai>=0.5.0",
    "groq>=0.4.0",
    "huggingface-hub>=0.23.0",
    "kaggle>=1.5.12",
    "langchain>=0.1.0",
    "langchain-community>=0.0.31",
    "langchain-huggingface>=0.2.0",
    "matplotlib>=3.8.0",
    "mypy>=1.5.0",
    "mysql-connector-python>=8.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pathlib>=1.0.1",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0",
    "pytest>=8.1.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=2.2.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.33.0",
    "tavily-python>=0.3.0",
    "tensorflow>=2.15.0",
    "tf-keras>=2.19.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.40.0",
    "uvicorn[standard]>=0.29.0",
]
